Encode username and password in the database URI
The username and password in the database URI may contain characters
that break (has special meaning). If so, they should be encoded, Why?
Because any operation done by Mongoose (like parsing or splitting the
URI) may result in error since the URI contains characters that has
special meaning.

Therefore we need to encode the username an password in the URI and send
URI to Mongoose, which in turn will decode both the username and
password (so Mongoose expects them to be encoded already) right before
sending the URI to the MongoDb server.

- Issue: https://github.com/aheckmann/muri/issues/9
- replica set:
https://docs.mongodb.com/manual/reference/glossary/#term-replica-set